botTon_fixed_5/botTon_fixed/requirements.txt
aiogram>=3.4.1
aiosqlite>=0.20.0
aiohttp>=3.9.3
python-dotenv>=1.0.1
pytonconnect>=0.2.2
aiogram-tonconnect>=0.1.0
qrcode[pil]>=7.4.2
tonsdk>=1.0.15
tonutils>=0.5.6

pyrogram>=2.0.106
tgcrypto>=1.2.3
aiohttp_cors>=0.7.0
lxml>=5.1.0
